Refactorings of Auth Extension
- Rename old PermissionService in backup to ControllerPermissionService
- Rename UserAlreadyExistException to UserLoginAlreadyExistException
- Remove Role:shortDescription
- Include anonymous/user role handling in login/logout
  - Add RoleService::installDefaultRoles
- Add update methods to RoleService & UserService

+++ Engine/Auth/Services/RoleService.php
@@ -4,7 +4,9 @@ namespace Oforge\Engine\Auth\Services;
 
 use Doctrine\ORM\EntityNotFoundException;
 use Doctrine\ORM\ORMException;
+use Oforge\Engine\Auth\AuthConstants;
 use Oforge\Engine\Auth\Exceptions\Role\RoleAlreadyExistException;
+use Oforge\Engine\Auth\Exceptions\Role\RoleImmutableException;
 use Oforge\Engine\Auth\Models\Role;
 use Oforge\Engine\Core\Abstracts\AbstractDatabaseAccess;
 use Oforge\Engine\Core\Manager\Events\Event;
@@ -22,6 +24,21 @@ class RoleService extends AbstractDatabaseAccess {
     }
 
     /**
+     *
+     */
+    public function installDefaultRoles() {
+        foreach ([AuthConstants::ROLE_ANONYMOUS, AuthConstants::ROLE_USER] as $roleName) {
+            try {
+                $this->create($roleName, true);
+            } catch (RoleAlreadyExistException $exception) {
+                // nothing to do
+            } catch (ORMException $exception) {
+                Oforge()->Logger()->logException($exception);
+            }
+        }
+    }
+
+    /**
      * @param string $name
      * @param string|null $shortDescription
      * @param bool $active
@@ -30,15 +47,14 @@ class RoleService extends AbstractDatabaseAccess {
      * @throws ORMException
      * @throws RoleAlreadyExistException
      */
-    public function create(string $name, ?string $shortDescription, bool $active = false) {
+    public function create(string $name, bool $active = false) {
         $role = $this->getByName($name);
         if ($role === null) {
             throw new RoleAlreadyExistException($name);
         }
         $role = Role::create([
-            'name'             => $name,
-            'active'           => $active,
-            'shortDescription' => $shortDescription,
+            'name'   => $name,
+            'active' => $active,
         ]);
 
         $this->entityManager()->create($role);
@@ -46,28 +62,6 @@ class RoleService extends AbstractDatabaseAccess {
         Oforge()->Events()->trigger(Event::create(Role::class . '::created', $roleData));
 
         return $roleData;
-    }
-
-    /**
-     * @param int $roleId
-     * @param bool $active
-     *
-     * @throws EntityNotFoundException
-     * @throws ORMException
-     */
-    public function setActivationById(int $roleId, bool $active) : void {
-        $this->setActivation($this->getById($roleId), $active, 'id', $roleId);
-    }
-
-    /**
-     * @param int $name
-     * @param bool $active
-     *
-     * @throws EntityNotFoundException
-     * @throws ORMException
-     */
-    public function setActivationByName(int $name, bool $active) : void {
-        $this->setActivation($this->getByName($name), $active, 'name', $name);
     }
 
     /**
@@ -104,6 +98,7 @@ class RoleService extends AbstractDatabaseAccess {
      * @param int $roleId
      *
      * @throws EntityNotFoundException
+     * @throws RoleImmutableException
      * @throws ORMException
      */
     public function removeById(int $roleId) : void {
@@ -114,6 +109,7 @@ class RoleService extends AbstractDatabaseAccess {
      * @param string $name
      *
      * @throws EntityNotFoundException
+     * @throws RoleImmutableException
      * @throws ORMException
      */
     public function removeByName(string $name) : void {
@@ -121,37 +117,95 @@ class RoleService extends AbstractDatabaseAccess {
     }
 
     /**
-     * @param Role|null $role
+     * @param int $roleId
      * @param bool $active
-     * @param string $key
-     * @param int|string $value
      *
      * @throws EntityNotFoundException
+     * @throws RoleImmutableException
      * @throws ORMException
      */
-    protected function setActivation(?Role $role, bool $active, string $key, $value) {
-        if ($role === null) {
-            throw EntityNotFoundException::fromClassNameAndIdentifier(Role::class, [$key => $value]);
-        }
-        $role->setActive($active);
-        $this->entityManager()->update($role);
+    public function setActivationById(int $roleId, bool $active) : void {
+        $this->update($this->getById($roleId), ['active' => $active], 'id', $roleId);
+    }
+
+    /**
+     * @param int $name
+     * @param bool $active
+     *
+     * @throws EntityNotFoundException
+     * @throws RoleImmutableException
+     * @throws ORMException
+     */
+    public function setActivationByName(int $name, bool $active) : void {
+        $this->update($this->getByName($name), ['active' => $active], 'name', $name);
+    }
+
+    /**
+     * @param int $roleId
+     * @param array $data
+     *
+     * @throws EntityNotFoundException
+     * @throws RoleImmutableException
+     * @throws ORMException
+     */
+    public function updateById(int $roleId, array $data) : void {
+        $this->update($this->getById($roleId), $data, 'id', $roleId);
+    }
+
+    /**
+     * @param string $name
+     * @param array $data
+     *
+     * @throws EntityNotFoundException
+     * @throws RoleImmutableException
+     * @throws ORMException
+     */
+    public function updateByName(string $name, array $data) : void {
+        $this->update($this->getByName($name), $data, 'name', $name);
     }
 
     /**
      * @param Role|null $role
      * @param string $key
-     * @param $value
+     * @param int|string $value
      *
      * @throws EntityNotFoundException
+     * @throws RoleImmutableException
      * @throws ORMException
      */
     protected function remove(?Role $role, string $key, $value) : void {
         if ($role === null) {
             throw EntityNotFoundException::fromClassNameAndIdentifier(Role::class, [$key => $value]);
         }
+        $roleName = $role->getName();
+        if (in_array($roleName, [AuthConstants::ROLE_ANONYMOUS, AuthConstants::ROLE_USER])) {
+            throw new RoleImmutableException("Role '$roleName' is not removable!");
+        }
         $roleData = $role->toArray();
         $this->entityManager()->remove($role);
         Oforge()->Events()->trigger(Event::create(Role::class . '::removed', $roleData));
+    }
+
+    /**
+     * @param Role|null $role
+     * @param array $data
+     * @param string $key
+     * @param int|string $value
+     *
+     * @throws EntityNotFoundException
+     * @throws RoleImmutableException
+     * @throws ORMException
+     */
+    protected function update(?Role $role, array $data, string $key, $value) : void {
+        if ($role === null) {
+            throw EntityNotFoundException::fromClassNameAndIdentifier(Role::class, [$key => $value]);
+        }
+        $roleName = $role->getName();
+        if (isset($data['active']) && in_array($roleName, [AuthConstants::ROLE_ANONYMOUS, AuthConstants::ROLE_USER])) {
+            throw new RoleImmutableException("Activation of role '$roleName' is not changeable!");
+        }
+        $role->fromArray($data);
+        $this->entityManager()->update($role);
     }
 
 }
